feat, refactor: added navigation from event card in events list to the start of registration flow. refactored the info page: separated the info page component to another page
